move the popup to caotains popup
