--- client/src/Components/Layouts/AddCaptain/index.jsx
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-import CollectionsPage from './addCaptain';
-
-// eslint-disable-next-line react/prefer-stateless-function
-class AddCaptain extends Component {
-  render() {
-    return (
-      <div>
-        <CollectionsPage />
-      </div>
-    );
-  }
-}
-
-export default AddCaptain;
